+++ CHANGELOG.md
@@ -50,6 +50,7 @@ All notable, unreleased changes to this project will be documented in this file.
   - Dry runs enabled: `AccountConfirmed`, `AccountConfirmationRequested`, `AccountChangeEmailRequested`
   - `AccountEmailChanged`, `AccountSetPasswordRequested`, `AccountDeleteRequested`, `GiftCardSent`,
   - `FulfillmentCreated`, `FulfillmentApproved`, `StaffSetPasswordRequested`
+- Add missing `FULFILLMENT_CREATED` event call to `automatically_fulfill_digital_lines_with_fulfillment_created` action. - #13823, by @Air-t
 
 ### Other changes
 
@@ -61,8 +62,9 @@ All notable, unreleased changes to this project will be documented in this file.
 - Add a new `product` field on `AssignedProductAttributeValue`. First part of a simplification of Attribute - Product relation from #12881. by @aniav
 - Lazy legacy webhooks payload generation - #13758 by @maarcingebala
 - Fix NoneType in `prodcutChannelsListingUpdate` - #13694 by @Manoj-gowra
+- Extended `AttributeValueTranslation.name` to 250 characters - #13776 by @aniav
 
-# 3.15.0 [Unreleased]
+# 3.15.0
 
 ### Breaking changes
 
@@ -148,7 +150,7 @@ Shipping methods can be removed by the user after it has been assigned to a chec
   - Add `metadata` to `AddressInput` field
 - Add support for unhandled field `Shop.trackInventoryByDefault` for `productVariantCreate` and `productVariantBulkCreate` - #13492 by @Smit-Parmar
 - Add `storedPaymentMethods` field to `User` and `Checkout` type - #13581
-- - Add `AttributeBulkUpdate` mutation - #13532 by @SzymJ
+- Add `AttributeBulkUpdate` mutation - #13532 by @SzymJ
 
 - Better checkout error feedback - #13458 by @korycins
   - Add `Checkout.problems` field